+++ NEWS.md
@@ -1,3 +1,10 @@
+# mortAAR 1.1
+- adding an option for plotting in color instead of linetype
+- adding more functions for checking representativity of age data, calculating reproduction indices, masculinity index, maternal mortality rate as well as corrected life tables
+- major re-factoring
+- rewriting of tests to work with testthat's new snapshot-functionality
+- adding two vignettes for life table correction and reproduction indices
+
 # mortAAR 1.0.2
 
 - tweaking the class-functions so that class mortaar_life_table is maintained when manipulating a mortaar_life_table_list